move function to helper file so i can test, plus remove unneeded permission checking function
